Queues ready for court notifications
