+++ apps/judicial-system/backend/src/app/modules/case/internalCase.service.ts
@@ -1,6 +1,7 @@
 import { Op } from 'sequelize'
 import CryptoJS from 'crypto-js'
 import { Sequelize } from 'sequelize-typescript'
+import format from 'date-fns/format'
 
 import {
   BadRequestException,
@@ -25,7 +26,7 @@ import {
 } from '@island.is/judicial-system/types'
 import type { User as TUser } from '@island.is/judicial-system/types'
 
-import { uuidFactory } from '../../factories'
+import { uuidFactory, nowFactory } from '../../factories'
 import {
   getCourtRecordPdfAsBuffer,
   getCourtRecordPdfAsString,
@@ -222,7 +223,7 @@ export class InternalCaseService {
       .then((pdf) => {
         const fileName = this.formatMessage(courtUpload.request, {
           caseType: caseTypes[theCase.type],
-          date: '',
+          date: ` ${format(nowFactory(), 'yyyy-MM-dd HH:mm')}`,
         })
 
         return this.courtService.createDocument(
