+++ apps/judicial-system/backend/src/app/modules/notification/models/sendNotification.response.ts
@@ -2,12 +2,7 @@ import { ApiProperty } from '@nestjs/swagger'
 
 import type { SendNotificationResponse as TSendNotificationResponse } from '@island.is/judicial-system/types'
 
-import { Notification } from './notification.model'
-
 export class SendNotificationResponse implements TSendNotificationResponse {
   @ApiProperty()
   notificationSent!: boolean
-
-  @ApiProperty({ type: Notification })
-  notification?: Notification
 }
