+++ apps/judicial-system/backend/src/app/modules/notification/notification.controller.ts
@@ -78,11 +78,15 @@ export class NotificationController {
       `Sending ${notification.type} notification for case ${caseId}`,
     )
 
-    if (notification.type === NotificationType.HEADS_UP) {
-      return this.notificationService
-        .addMessagesForHeadsUpNotificationToQueue(theCase)
-        .then(() => ({ notificationSent: true }))
-        .catch(() => ({ notificationSent: false }))
+    if (
+      [NotificationType.HEADS_UP, NotificationType.READY_FOR_COURT].includes(
+        notification.type,
+      )
+    ) {
+      return this.notificationService.addMessagesForNotificationToQueue(
+        theCase,
+        notification,
+      )
     }
 
     return this.notificationService.sendCaseNotification(
