+++ apps/judicial-system/message-handler/src/app/messageHandler.service.ts
@@ -106,6 +106,13 @@ export class MessageHandlerService implements OnModuleDestroy {
           { type: NotificationType.HEADS_UP },
         )
         break
+      case MessageType.SEND_READY_FOR_COURT_NOTIFICATION:
+        handled = await this.internalDeliveryService.deliver(
+          message.caseId,
+          'notification',
+          { type: NotificationType.READY_FOR_COURT },
+        )
+        break
       case MessageType.SEND_DEFENDANTS_NOT_UPDATED_AT_COURT_NOTIFICATION:
         handled = await this.internalDeliveryService.deliver(
           message.caseId,
