+++ apps/judicial-system/message-handler/src/app/test/handleMessage.spec.ts
@@ -323,6 +323,32 @@ describe('MessageHandlerService - Handle message', () => {
     })
   })
 
+  describe('send ready for court notification', () => {
+    let then: Then
+
+    beforeEach(async () => {
+      then = await givenWhenThen({
+        type: MessageType.SEND_READY_FOR_COURT_NOTIFICATION,
+        caseId,
+      })
+    })
+
+    it('should send a ready for court notification', async () => {
+      expect(fetch).toHaveBeenCalledWith(
+        `${config.backendUrl}/api/internal/case/${caseId}/notification`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            authorization: `Bearer ${config.backendAccessToken}`,
+          },
+          body: JSON.stringify({ type: NotificationType.READY_FOR_COURT }),
+        },
+      )
+      expect(then.result).toBe(true)
+    })
+  })
+
   describe('send defendants not updated at court notification', () => {
     let then: Then
