+++ libs/judicial-system/message/src/lib/message.ts
@@ -9,6 +9,7 @@ export enum MessageType {
   DELIVER_CASE_TO_POLICE = 'DELIVER_CASE_TO_POLICE',
   ARCHIVE_CASE_FILE = 'ARCHIVE_CASE_FILE',
   SEND_HEADS_UP_NOTIFICATION = 'SEND_HEADS_UP_NOTIFICATION',
+  SEND_READY_FOR_COURT_NOTIFICATION = 'SEND_READY_FOR_COURT_NOTIFICATION',
   SEND_DEFENDANTS_NOT_UPDATED_AT_COURT_NOTIFICATION = 'SEND_DEFENDANTS_NOT_UPDATED_AT_COURT_NOTIFICATION',
   SEND_RULING_NOTIFICATION = 'SEND_RULING_NOTIFICATION',
 }
